requirements.txt
asgiref==3.6.0
Django==4.1.5
sqlparse==0.4.3
psycopg2-binary==2.9.5
django-phonenumber-field==7.1.0
django-smalluuid==1.2.1